+++ CHANGELOG.md
@@ -3,3 +3,5 @@
         * **IMPROVEMENT** :-1: : Cleaned `utils`.
     * **JSON**:
         * **IMPROVEMENT** :raised_hands: : Extract Jackson folder, removed converter service.
+    * **REPO**:
+        * **IMPROVEMENT** :raised_hands: : Migrated in memory external repos.